+++ src/httprpc.cpp
@@ -3,8 +3,8 @@
 #include "chainparams.h"
 #include "httpserver.h"
 #include "key_io.h"
-#include "rpcprotocol.h"
-#include "rpcserver.h"
+#include "rpc/protocol.h"
+#include "rpc/server.h"
 #include "random.h"
 #include "sync.h"
 #include "util.h"
