+++ src/rpc/blockchain.cpp
@@ -10,7 +10,8 @@
 #include "consensus/validation.h"
 #include "main.h"
 #include "primitives/transaction.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
+#include "streams.h"
 #include "sync.h"
 #include "util.h"
