+++ src/wallet/rpcdisclosure.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "init.h"
 #include "key_io.h"
 #include "main.h"
